moved MockConstitutiveLaw in a separate file and removed StubConstitutiveLaw
